+++ squire_lib/Cargo.toml
@@ -9,8 +9,8 @@ authors = ["TylerBloom <tylerbloom2222@gmail.com>"]
 crate-type = ["lib", "staticlib"]
 
 [features]
-default = ["deck_sites"]
-#default = ["ffi", "mtgjson/deck_sites"]
+#default = ["deck_sites"]
+default = ["ffi", "mtgjson/deck_sites"]
 deck_sites = ["mtgjson/deck_sites"]
 ffi = ["cbindgen", "libc"]
 wasm = ["chrono/wasmbind"]
