+++ squire_lib/ffi_build.bat
