+++ squire_lib/ffi_build.sh
