+++ squire_lib/src/pairings/mod.rs
@@ -9,8 +9,8 @@ use crate::{
     error::TournamentError,
     identifiers::PlayerId,
     operations::{OpData, OpResult},
-    player::PlayerRegistry,
-    round::RoundRegistry,
+    players::PlayerRegistry,
+    rounds::RoundRegistry,
     scoring::{Score, Standings},
     settings::PairingSetting,
     tournament::TournamentPreset,
