+++ squire_lib/tests/swiss_pairings.rs
@@ -7,8 +7,8 @@ mod tests {
         accounts::{SharingPermissions, SquireAccount},
         identifiers::{PlayerId, UserAccountId},
         pairings::PairingSystem,
-        player::PlayerRegistry,
-        round::{RoundRegistry, RoundResult},
+        players::PlayerRegistry,
+        rounds::{RoundRegistry, RoundResult},
         settings::SwissPairingsSetting,
         scoring::StandardScoring,
         tournament::TournamentPreset,
