+++ squire_lib/tests/utils.rs
@@ -1,4 +1,4 @@
-use squire_lib::player::Player;
+use squire_lib::players::Player;
 
 pub fn spoof_player() -> Player {
     Player::new(uuid::Uuid::new_v4().to_string())
